Remove terraform-output artifact from E2E tests
The terraform output was being captured to files and uploaded as an
artifact, but this adds complexity without much benefit since terraform
output is already visible in the CI logs.
