+++ .github/workflows/e2e_test.yml
@@ -51,8 +51,6 @@ jobs:
       PREEMPTIBLE_NODES: ${{ vars.E2E_TEST_PREEMPTIBLE_NODES }}
       SSH_KEYS: ${{ vars.E2E_TEST_SSH_KEYS }}
       O11Y_ACCESS_TOKEN: ${{ secrets.E2E_O11Y_ACCESS_TOKEN }}
-      OUTPUT_LOG_FILE: "${{ github.workspace }}/output/output.log"
-      OUTPUT_ERR_FILE: "${{ github.workspace }}/output/output.err"
 
     steps:
       - name: Checkout repository
@@ -413,14 +411,3 @@ jobs:
               echo "- [$hash]($terraform_repo_url/commit/$hash): $msg ($date) <$author>"
             done
           } >> $GITHUB_STEP_SUMMARY
-
-      - name: Add errors output to job summary
-        if: ${{ always() }}
-        run: cat $OUTPUT_ERR_FILE >> $GITHUB_STEP_SUMMARY
-
-      - name: Upload terraform output
-        if: ${{ always() }}
-        uses: actions/upload-artifact@v6
-        with:
-          name: terraform-output
-          path: "${{ github.workspace }}/output"
