+++ test/e2e/e2e_test.go
@@ -3,6 +3,9 @@
 package e2e_test
 
 import (
+	"fmt"
+	"os"
+	"strings"
 	"testing"
 
 	"github.com/gruntwork-io/terratest/modules/terraform"
@@ -20,20 +23,14 @@ func TestTerraformApply(t *testing.T) {
 
 	commonOptions := setupTerraformOptions(t, cfg)
 
-	ensureOutputFiles(t, cfg)
-
 	terraform.Init(t, &commonOptions)
 	terraform.WorkspaceSelectOrNew(t, &commonOptions, "e2e-test")
 
 	// Pre-test cleanup to ensure clean state
-	output, err := terraform.DestroyE(t, &commonOptions)
-	writeOutputs(t, cfg, "TestTerraformApply", "destroy", output, err)
-	require.NoError(t, err)
+	terraform.Destroy(t, &commonOptions)
 
 	// Apply terraform configuration
-	output, err = terraform.ApplyE(t, &commonOptions)
-	writeOutputs(t, cfg, "TestTerraformApply", "apply", output, err)
-	require.NoError(t, err)
+	terraform.Apply(t, &commonOptions)
 
 	// Note: No defer destroy - cleanup is handled by TestTerraformDestroy
 	// This allows cluster state collection on failure in CI
@@ -49,13 +46,29 @@ func TestTerraformDestroy(t *testing.T) {
 
 	commonOptions := setupTerraformOptions(t, cfg)
 
-	ensureOutputFiles(t, cfg)
-
 	terraform.Init(t, &commonOptions)
 	terraform.WorkspaceSelectOrNew(t, &commonOptions, "e2e-test")
 
 	// Destroy the infrastructure
-	output, err := terraform.DestroyE(t, &commonOptions)
-	writeOutputs(t, cfg, "TestTerraformDestroy", "destroy", output, err)
-	require.NoError(t, err)
+	terraform.Destroy(t, &commonOptions)
+}
+
+func setupTerraformOptions(t *testing.T, cfg testConfig) terraform.Options {
+	tfVars := readTFVars(t, fmt.Sprintf("%s/terraform.tfvars", cfg.PathToInstallation))
+	tfVars = overrideTestValues(t, tfVars, cfg)
+
+	envVarsList := os.Environ()
+	envVars := make(map[string]string)
+	for _, envVar := range envVarsList {
+		pair := strings.SplitN(envVar, "=", 2)
+		require.Len(t, pair, 2)
+		envVars[pair[0]] = pair[1]
+	}
+
+	return terraform.Options{
+		TerraformDir: cfg.PathToInstallation,
+		Vars:         tfVars,
+		EnvVars:      envVars,
+		NoColor:      true,
+	}
 }
